perf: streamline CI preparation and test sharding
