+++ package.json
@@ -44,7 +44,7 @@
 		"setup": "bun ./scripts/install-frozen.mjs && bun ./scripts/install-frozen.mjs ui && bun run setup-contracts && bun run ui:vendor && cd ui && bun run refresh:shared-dependency && tsc && bun run build:workers && bun ./build/tests.mts",
 		"setup-contracts": "cd solidity && bun run setup",
 		"compile-contracts": "cd solidity && bun run compile-contracts",
-		"generate": "bun run generate:contracts && bun run generate:ui",
+		"generate": "bun run generate:contracts && bun run generate:ui-vendor",
 		"check:mainnet-deployment": "bun run generate && bun run refresh:shared-dependencies && cd ui && bun x tsc --project tsconfig.json && cd .. && bun ./scripts/check-mainnet-deployment.mts",
 		"check:mainnet-deployment:generated": "bun run generate && bun run refresh:shared-dependencies && bun ./scripts/check-mainnet-deployment.mts",
 		"check:mainnet-deployment:current": "bun ./scripts/check-mainnet-deployment.mts",
@@ -53,12 +53,14 @@
 		"tsc:solidity": "cd solidity && bun ../scripts/ensure-shared-package-fresh.mjs --refresh && bun x tsc --project tsconfig.typecheck.json",
 		"tsc:solidity:current": "cd solidity && bun x tsc --project tsconfig.typecheck.json",
 		"tsc:ci:current": "bun run tsc:app && bun run tsc:scripts && bun run tsc:solidity:current",
+		"ci:preflight:current": "bun ./scripts/run-ci-preflight.mts",
 		"tsc": "bun run ensure-shared-build && bun run check:shared-dependencies && bun run tsc:app && bun run tsc:scripts && bun run tsc:solidity",
 		"ui:setup": "bun ./scripts/install-frozen.mjs && bun ./scripts/install-frozen.mjs ui && cd solidity && bun run setup && cd .. && bun run ui:vendor && cd ui && bun run refresh:shared-dependency && tsc && bun run build:workers && bun ./build/tests.mts",
 		"app:build": "bun run generate && cd ui && bun run build && bun run build:tests",
 		"ui:build": "bun run app:build",
 		"ui:build:prod": "bun run check:mainnet-deployment && cd ui && bun run build:prod",
 		"ui:build:prod:current": "bun run check:mainnet-deployment:current && cd ui && bun ./build/production.mts",
+		"ui:build:prod:prepared": "bun run refresh:shared-dependencies && cd ui && bun x tsc --project tsconfig.json && cd .. && bun run ui:build:prod:current",
 		"ui:workers": "cd ui && bun ./build/workers.mts",
 		"app:watch": "bun run app:build && bun ./ui/build/watch.mts",
 		"ui:watch": "bun run app:watch",
@@ -93,7 +95,11 @@
 		"docs:check-reference-values": "bun ./scripts/check-docs-reference-values.mts",
 		"docs:generate-contract-reference": "bun ./scripts/generate-contract-interaction-reference.mts",
 		"docs:check": "bun run docs:check-html && bun run docs:check-examples && bun run docs:check-content && bun run docs:check-contract-reference && bun run docs:check-reference-values && bun run check:mainnet-deployment:generated",
-		"check": "bunx @biomejs/biome check package.json .prettierrc.json shared/package.json ui/package.json solidity/package.json tsconfig.scripts.json bun-test-setup*.ts README.md AGENTS.md .codex/review-contract.md docs scripts shared/ts solidity/ts ui/AGENTS.md ui/ts ui/build ui/dev-server.ts ui/css && bun scripts/lint-no-bare-catch.mts && bun scripts/lint-no-signal-wrapper-comparisons.mts && bun scripts/lint-no-direct-viem-imports.mts && bun scripts/lint-no-nested-solidity-ternaries.mts && bun run lint:ui-layer-boundaries && bun run docs:check && bun run lint:ui-strings && bun run check:agents && bun run check:solidity-format",
+		"docs:check:current": "bun run docs:check-html && bun run docs:check-examples && bun run docs:check-content && bun run docs:check-contract-reference && bun run docs:check-reference-values && bun run check:mainnet-deployment:current",
+		"check:static": "bunx @biomejs/biome check package.json .prettierrc.json shared/package.json ui/package.json solidity/package.json tsconfig.scripts.json bun-test-setup*.ts README.md AGENTS.md .codex/review-contract.md docs scripts shared/ts solidity/ts ui/AGENTS.md ui/ts ui/build ui/dev-server.ts ui/css && bun scripts/lint-no-bare-catch.mts && bun scripts/lint-no-signal-wrapper-comparisons.mts && bun scripts/lint-no-direct-viem-imports.mts && bun scripts/lint-no-nested-solidity-ternaries.mts && bun run lint:ui-layer-boundaries",
+		"check:repository": "bun run lint:ui-strings && bun run check:agents && bun run check:solidity-format",
+		"check": "bun run check:static && bun run docs:check && bun run check:repository",
+		"check:ci:current": "bun run check:static && bun run docs:check:current && bun run check:repository",
 		"check:changed": "bun scripts/check-changed.mts && bun run lint:ui-strings && bun run check:agents",
 		"check:agents": "bun scripts/check-agent-configs.mts",
 		"check:solidity-format": "bun x prettier --check 'solidity/contracts/**/*.sol'",
