+++ ui/ts/tests/simulation/activeEnvironment.test.ts
@@ -2,46 +2,21 @@
 
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, test } from 'bun:test'
 import { getAddress } from '@zoltar/shared/ethereum'
-import { loadAllSecurityPools, loadDeploymentStatusOracleSnapshot, loadErc20Balance, loadSecurityVaultDetails, loadZoltarUniverseSummary } from '../../protocol/index.js'
-import { getDeploymentSteps } from '../../protocol/deployment.js'
+import { loadDeploymentStatusOracleSnapshot, loadErc20Balance } from '../../protocol/index.js'
 import { getWalletScopedAccountAddress, getWrongNetworkMessage, isSupportedAppChain } from '../../lib/network.js'
 import { getActiveBackend, initializeActiveEnvironment, installActiveEnvironmentForTesting, resetActiveEnvironmentForTesting, shouldUseSimulationLocation } from '../../lib/activeEnvironment.js'
 import { SIMULATION_BLOCK_INTERVAL_SECONDS, SIMULATION_INITIAL_TIMESTAMP } from '../../simulation/clock.js'
 import { parseSavedSimulationStateEnvelope, persistSavedSimulationState, serializeSavedSimulationStateEnvelope } from '../../simulation/savedStates.js'
 import { createSimulationBackend } from '../../simulation/tevmBackend.js'
-import type { SimulationScenario } from '../../simulation/scenarios.js'
 import { createFakeBackend, createFakeSimulationProfile } from '../testUtils/fakeBackend.js'
 import { installDomEnvironment } from '../testUtils/domEnvironment.js'
+import { createBootstrappedSimulationBackendWithRetry, resetSelectedAccountAndTransactionDelay, type SimulationBackend } from './testUtils.js'
 
 const DEFAULT_SIMULATION_REP_PER_ETH_PRICE = 3n * 10n ** 18n
 const SIMULATION_REP_MINT_AMOUNT = 1_000_000n * 10n ** 18n
-const SEEDED_REP_DEPOSIT = 10_000n * 10n ** 18n
-const SEEDED_SECURITY_BOND_ALLOWANCE = 100n * 10n ** 18n
-type SimulationBackend = Awaited<ReturnType<typeof createSimulationBackend>>
 afterEach(() => {
 	resetActiveEnvironmentForTesting()
 })
-
-async function createBootstrappedSimulationBackendWithRetry(scenario: SimulationScenario, maxAttempts = 2) {
-	let lastError: unknown = undefined
-	for (let attempt = 1; attempt <= maxAttempts; attempt += 1) {
-		const backend = await createSimulationBackend({ scenario })
-		try {
-			await backend.bootstrap()
-			return backend
-		} catch (error) {
-			lastError = error
-			await backend.dispose()
-		}
-	}
-	throw lastError instanceof Error ? lastError : new Error(`Failed to bootstrap ${scenario} simulation backend`)
-}
-
-async function resetSelectedAccountAndTransactionDelay(backend: SimulationBackend) {
-	const primaryAccount = backend.accounts[0]
-	if (primaryAccount !== undefined && backend.selectedAccount !== primaryAccount) await backend.selectAccount(primaryAccount)
-	backend.setTransactionDelayMilliseconds(0)
-}
 
 void describe('active environment', () => {
 	void test('uses the injected backend by default when no environment has been initialized', () => {
@@ -193,45 +168,21 @@ void describe('active environment', () => {
 void describe('simulation backend', () => {
 	let coldBaselineBackend: SimulationBackend
 	let warmBaselineBackend: SimulationBackend
-	let deployedBackend: SimulationBackend
-	let securityPoolBackend: SimulationBackend
-	let securityPoolX2Backend: SimulationBackend
-	let securityPoolX2AuctionBackend: SimulationBackend
 
 	beforeAll(async () => {
 		coldBaselineBackend = await createSimulationBackend({ scenario: 'baseline' })
-		const [nextWarmBaselineBackend, nextDeployedBackend, nextSecurityPoolBackend, nextSecurityPoolX2Backend, nextSecurityPoolX2AuctionBackend] = await Promise.all([
-			createBootstrappedSimulationBackendWithRetry('baseline'),
-			createBootstrappedSimulationBackendWithRetry('deployed'),
-			createBootstrappedSimulationBackendWithRetry('security-pool', 1),
-			createBootstrappedSimulationBackendWithRetry('securitypoolx2', 1),
-			createBootstrappedSimulationBackendWithRetry('securitypoolx2-auction', 1),
-		])
-		warmBaselineBackend = nextWarmBaselineBackend
-		deployedBackend = nextDeployedBackend
-		securityPoolBackend = nextSecurityPoolBackend
-		securityPoolX2Backend = nextSecurityPoolX2Backend
-		securityPoolX2AuctionBackend = nextSecurityPoolX2AuctionBackend
+		warmBaselineBackend = await createBootstrappedSimulationBackendWithRetry('baseline')
 		warmBaselineBackend.setTransactionDelayMilliseconds(0)
-		deployedBackend.setTransactionDelayMilliseconds(0)
-		securityPoolBackend.setTransactionDelayMilliseconds(0)
-		securityPoolX2Backend.setTransactionDelayMilliseconds(0)
-		securityPoolX2AuctionBackend.setTransactionDelayMilliseconds(0)
 	}, 180_000)
 
 	beforeEach(async () => {
 		await resetSelectedAccountAndTransactionDelay(coldBaselineBackend)
 		await resetSelectedAccountAndTransactionDelay(warmBaselineBackend)
-		await resetSelectedAccountAndTransactionDelay(deployedBackend)
 	}, 30_000)
 
 	afterAll(async () => {
 		if (coldBaselineBackend !== undefined) await coldBaselineBackend.dispose()
 		if (warmBaselineBackend !== undefined) await warmBaselineBackend.dispose()
-		if (deployedBackend !== undefined) await deployedBackend.dispose()
-		if (securityPoolBackend !== undefined) await securityPoolBackend.dispose()
-		if (securityPoolX2Backend !== undefined) await securityPoolX2Backend.dispose()
-		if (securityPoolX2AuctionBackend !== undefined) await securityPoolX2AuctionBackend.dispose()
 	}, 30_000)
 
 	void test('reports wallet presence and returns the selected account', async () => {
@@ -365,31 +316,6 @@ void describe('simulation backend', () => {
 			await backend.dispose()
 		}
 	}, 30_000)
-
-	void test('keeps the deployed-scenario fork threshold in sync after minting REP', async () => {
-		const backend = deployedBackend
-
-		const readClient = backend.createReadClient()
-		const zoltarStep = getDeploymentSteps().find(step => step.id === 'zoltar')
-		if (zoltarStep === undefined) throw new Error('Expected the Zoltar deployment step')
-		const universeSummaryBefore = await loadZoltarUniverseSummary(readClient, 0n)
-		if (universeSummaryBefore === undefined) {
-			throw new Error('Expected the genesis Zoltar universe to be available in the deployed scenario')
-		}
-		const zoltarBalanceBefore = await readClient.getBalance({ address: zoltarStep.address })
-
-		await backend.mintRep(SIMULATION_REP_MINT_AMOUNT)
-
-		const universeSummaryAfter = await loadZoltarUniverseSummary(readClient, 0n)
-		if (universeSummaryAfter === undefined) {
-			throw new Error('Expected the genesis Zoltar universe after minting REP')
-		}
-		const zoltarBalanceAfter = await readClient.getBalance({ address: zoltarStep.address })
-
-		expect(universeSummaryAfter.totalTheoreticalSupply).toBe(universeSummaryBefore.totalTheoreticalSupply + SIMULATION_REP_MINT_AMOUNT)
-		expect(universeSummaryAfter.forkThreshold).toBe(universeSummaryBefore.forkThreshold + SIMULATION_REP_MINT_AMOUNT / 20n)
-		expect(zoltarBalanceAfter).toBe(zoltarBalanceBefore)
-	}, 60_000)
 
 	void test('submits simulation writes without deprecated Tevm transaction RPC warnings', async () => {
 		const backend = await createSimulationBackend({ scenario: 'baseline' })
@@ -534,50 +460,5 @@ void describe('simulation backend', () => {
 		} finally {
 			await sourceBackend.dispose()
 		}
-	}, 60_000)
-
-	void test('bootstraps the deployed scenario with app contracts already deployed', async () => {
-		const backend = deployedBackend
-
-		const deploymentSnapshot = await loadDeploymentStatusOracleSnapshot(backend.createReadClient())
-
-		expect(backend.currentScenario).toBe('deployed')
-		expect(deploymentSnapshot.augurPlaceHolderDeployed).toBe(true)
-		expect(deploymentSnapshot.deploymentStatuses.every(step => step.deployed)).toBe(true)
-	}, 30_000)
-
-	void test('bootstraps seeded security-pool scenarios without reverting', async () => {
-		const seededScenarios = [
-			{ backend: securityPoolBackend, scenario: 'security-pool' },
-			{ backend: securityPoolX2Backend, scenario: 'securitypoolx2' },
-			{ backend: securityPoolX2AuctionBackend, scenario: 'securitypoolx2-auction' },
-		] satisfies Array<{ backend: SimulationBackend; scenario: SimulationScenario }>
-		for (const { backend, scenario } of seededScenarios) {
-			expect(backend.currentScenario).toBe(scenario)
-			const readClient = backend.createReadClient()
-			const pools = await loadAllSecurityPools(readClient)
-			expect(pools.length).toBeGreaterThan(0)
-		}
-	}, 180_000)
-
-	void test('bootstraps the security-pool scenario with one undercollateralized seeded vault', async () => {
-		const backend = securityPoolBackend
-		const primaryAccount = backend.accounts[0]
-		if (primaryAccount === undefined) throw new Error('Expected seeded simulation QA accounts')
-
-		const readClient = backend.createReadClient()
-		const pools = await loadAllSecurityPools(readClient)
-		const seededPool = pools[0]
-		if (seededPool === undefined) throw new Error('Expected a seeded security pool')
-		const seededVault = await loadSecurityVaultDetails(readClient, seededPool.securityPoolAddress, primaryAccount)
-		if (seededVault === undefined) throw new Error('Expected a seeded security vault')
-
-		expect(backend.currentScenario).toBe('security-pool')
-		expect(pools).toHaveLength(1)
-		expect(seededPool.vaultCount).toBe(1n)
-		expect(seededPool.totalRepDeposit).toBe(SEEDED_REP_DEPOSIT)
-		expect(seededPool.totalSecurityBondAllowance).toBe(SEEDED_SECURITY_BOND_ALLOWANCE)
-		expect(seededVault.repDepositShare).toBe(SEEDED_REP_DEPOSIT)
-		expect(seededVault.securityBondAllowance).toBe(SEEDED_SECURITY_BOND_ALLOWANCE)
 	}, 60_000)
 })
